refactor(circuit): consolidate cross-mixin TYPE_CHECKING blocks into CircuitProtocol
The five mixins under circuit_internals (LifecycleMixin, SubsystemTreeMixin,
HamiltonianAssemblyMixin, CircuitSymMethods, CircuitPlot) each declared a
parallel set of cross-mixin attribute and method types inside their own
``if TYPE_CHECKING:`` blocks. The same attribute (e.g. ``var_categories``)
was redeclared with subtly different types across files; adding a new
shared attribute required updating every mixin, and a contributor changing
the type in one mixin had no automatic check against the others.

Adds circuit_internals/_protocols.py defining ``CircuitProtocol`` — the
single source of truth for the cross-mixin attribute / method surface.
At runtime the class is empty (its body is gated under TYPE_CHECKING);
mypy sees the full ~60 attribute and ~25 method declarations. Each
mixin now inherits from CircuitProtocol unconditionally:

    class LifecycleMixin(ABC, CircuitProtocol): ...
    class SubsystemTreeMixin(ABC, CircuitProtocol): ...
    class HamiltonianAssemblyMixin(ABC, CircuitProtocol): ...
    class CircuitSymMethods(ABC, CircuitProtocol): ...
    class CircuitPlot(ABC, CircuitProtocol): ...

Net: -295 lines (the union of removed in-class declarations) + 205 lines
(the protocol). Runtime MRO grows by one class (CircuitProtocol appears
once via Python's deduplication, not five times — verified via
Circuit.__mro__).

One narrow type was kept as ``Any``: ``hilbert_space``. Surfacing the
precise ``HilbertSpace`` type exposed three latent typing bugs at call
sites like ``self.hilbert_space["bare_evals"]`` that pre-date this
refactor; addressing those is a separate workstream. A comment in
_protocols.py records the deferred tightening.

Tests: 54 circuit + characterization tests pass; mypy clean on the
six modified files.
